Rename "Wakeable" modules to "Thenable"
The distinction between "wakeable" and "thenable" isn't really
important in these files anymore so we should just pick one.
