+++ packages/react-server/src/ReactFizzHooks.js
@@ -21,7 +21,7 @@ import type {
 
 import type {ResponseState} from './ReactServerFormatConfig';
 import type {Task} from './ReactFizzServer';
-import type {ThenableState} from './ReactFizzWakeable';
+import type {ThenableState} from './ReactFizzThenable';
 
 import {readContext as readContextImpl} from './ReactFizzNewContext';
 import {getTreeId} from './ReactFizzTreeContext';
@@ -29,7 +29,7 @@ import {
   getPreviouslyUsedThenableAtIndex,
   createThenableState,
   trackUsedThenable,
-} from './ReactFizzWakeable';
+} from './ReactFizzThenable';
 
 import {makeId} from './ReactServerFormatConfig';
